// console/src/api/generated/models/link-verification-trigger-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Result returned after link verification work is accepted.
 * @export
 * @interface LinkVerificationTriggerResult
 */
export interface LinkVerificationTriggerResult {
    /**
     * Number of links accepted for background verification.
     * @type {number}
     * @memberof LinkVerificationTriggerResult
     */
    'acceptedCount'?: number;
    /**
     * Link names accepted for background verification.
     * @type {Array<string>}
     * @memberof LinkVerificationTriggerResult
     */
    'acceptedNames'?: Array<string>;
    /**
     * Number of requested links already running verification.
     * @type {number}
     * @memberof LinkVerificationTriggerResult
     */
    'alreadyRunningCount'?: number;
    /**
     * Link names skipped because verification is already running.
     * @type {Array<string>}
     * @memberof LinkVerificationTriggerResult
     */
    'alreadyRunningNames'?: Array<string>;
    /**
     * Number of requested links skipped because they could not be resolved.
     * @type {number}
     * @memberof LinkVerificationTriggerResult
     */
    'skippedCount'?: number;
    /**
     * Requested link names skipped because they could not be resolved.
     * @type {Array<string>}
     * @memberof LinkVerificationTriggerResult
     */
    'skippedNames'?: Array<string>;
}